+++ source/Images/WebPImage.cpp
@@ -25,7 +25,7 @@ namespace Images
 			&width, &height
 		);
 		
-		_size = {width, height, 1};
+		_size = {static_cast<std::size_t>(width), static_cast<std::size_t>(height), 1};
 	}
 	
 	WebPImage::~WebPImage()
